Move measureTextWidths to render/canvas
